Move key upload to a separate page.
The upload page explains how to generate an SSL key and hwo to extract
the public key for upload. Changed the schema to include the file name
of the key for a reminder to the user. Remove the display of the key
prefix because it makes no sense with ssl keys.