Use the sha1 hash of the slice urn for ABAC.
In order to generate ABAC slice-related credentials let's try using
the 40-character sha1 hash of the slice urn. The urn is already in use
in GENI as the AM API argument to denote the slice, so hashing it
makes an easy (and ABAC-compatible) value for checking privileges in
the RT1-Lite world (Privilege_Object). This is easy to change as our
thinking evolves.