Catch missing eppn from idp.
The eduPersonPrincipalName (eppn) is a required attribute. It is used
as the persistent key to link look up an existing identity. If an IdP
is not sharing it, the GENI portal can't really work. There are no
other guaranteed fields that uniquely identify an individual. If the
IdP is not sharing eppn, put up a custom error message to explain the
problem to the user.