Use a non-existent hostname for the authority URL.
We were getting a lot of "slice is busy errors from
pgeni3.gpolab.bbn.com when using the FQDN of internal lab VMs. My
theory is that those connections had to time out. With a non-existent
hostname, the connection should error immediately so we'll get fewer
"busy" errors. It remains to be seen whether this works in practices.