Add missing arg in internal add_project_member call.
Add a NULL $message on internal call to add_project_member within
create_project. Undo the $signer logic in add_project_member because
it is broken. The signer holds a cert/key, but in this case we have
not private key. The UUID is not sufficient. The PA probably has to
sign this call, and the CS will have to allow the PA to make
assertions.