Fix sliver expiration bug.
Fix the name of the date/time field to sliver_expiration so the renew
handler can find it. Also relabel the expiration column to be renew
instead of expiration to more accurately reflect the displayed
data. Finally, add a hidden field for the am_id so that the resources
can be renewed at a single aggregate.