Ticket 648: undefined variables in amstatus with am_id
Most of the errors on the ticket were fixed with 645. This fixes
instances where am_id is undefined. If am_id was not specified as
part of the URL, then $am_id was never defined when using the tool
tool-lookupids.php. Now $am_id is initialized as an empty string
prior to the include of tool-lookupids.php.