Ticket 655: Allow for download public SSH key
A new file, downloadsshpublickey.php, takes care of retrieving the
public ssh keys. It is a copy of downloadsshkey.php (used to download
private keys) and it replaces instances of 'private' with 'public'.
All of the user's public ssh keys (whether they were generated
from the portal or not) are available for download. This fix was
tested with having 1) a generated keypair and 2) a user-uploaded
public key and 3) both.