Allow openid users to activate their account.
If a new user comes to the portal the first time through openid,
properly redirect to the kmactivate page. Note that this will
eventually land the user at the home page, not back in the openid
context. The situation hasn't ever happened in the wild, so this fix
is good enough for now.