Fix a type juggling issue.
It is not possible to invoke a chapi xml-rpc endpoint with no
underscore in its name because the result of strpos is misinterpreted
in XMLRPCClient. The first character of the method without underscores
is accidentally dropped.

strpos($haystack, $needle) returns FALSE if $needle is not found, and
0 if $needle is found at position zero. These two values are
equivalent when type juggling is used. Distinguish between the two by
using the triple equals operator.