Remove obsolete files from Makefiles
I will update the Makefile when I remove a file.
I will update the Makefile when I remove a file.
I will update the Makefile when I remove a file.
I will update the Makefile when I remove a file.
I will update the Makefile when I remove a file.