Generate the agg_nick_cache from SR data
Instead of copying a stale aggregate nickname cache, generate one from
service registry data. Receive the short name from the SR by adding it
to the list of known keys to translate. Create a new function to write
the agg_nick_cache to file using INI format with the correct preamble
and all the aggregates who have a short name, URN, and URL. Call this
new function instead of copying the file.