Overhaul jfed.php
Handle missing or expired certificates. Instead of double fetching to
figure out what the issue was, inline the library code that is only
used in this file. This enables improved error messages.

Handle conditions in a careful order to minimize calls to the
clearinghouse. Use messages and links specific to the condition to
help the user fix the issue if possible.