Issue #1673: Make self asserting an email address be self-service. Adding new table km_confirm_email to hold the nonce and email for a given eppn.
kmnoemail.php will clear entries from that table that are 7 days old. If you shouldn't be here (no eppn, got email, already asserted email) it redirects. It reads any partially asserted email address to pre fill out a form asking for an email address, which goes to....
New page: kmsendemail.php redirects if you're in the wrong place, sends you back to kmnoemail on a bad email address. Otherwise it replaces any existing entry in km_no_email, and sends the experimenter email to confirm their address (BCCing admins), and telling them to look for this email.
The email points them to new page kmconfirmemail.php.
That page redirects if you're in the wrong place, complains if this is a bad link (id/nonce/eppn triple). Otherwise it puts the email/eppn in km_asserted_attributes and removes the entry from km_confirm_email.

To make this work properly, added a bunch to the KM CSS and a line to kmheader (to properly show/validate the form, and to support tool-showmessage and its clear button).

Note that this change adds a new DB table; run kmtool/db/postgresql/update-2.sql to create this table on an existing portal.