Add CentOS 7 httpd config for OpenID
Use the newer apache permission directives in a CentOS 7 apache
config to be included in the portal-ssl apache config file. Include
the new file in the portal-ssl template. Use the GNU replacement
commands to set paths in the new config file. Use the GNU replacement
commands to set paths in the existing apache config as well. Include
the new file in the openid Makefile and RPM spec file.