Use a placeholder ssh key in geni-sync-wireless
If a project lead has no ssh keys it disrupts the synchronization
with the wireless testbed(s). Users must have an ssh key. The old
solution of skipping these users causes the script to abort early
if a project lead lacks ssh keys.

After consultation with WinLab, we are using a placeholder ssh key
for users who do not have ssh keys in GENI. This allows the sync to
work and will make it obvious to users that something is amiss when
they look at their profile and see "missing GENI ssh key".