Prevent GPO users from transferring accounts
GPO user accounts cannot be the destination of an account transfer.
Prevent this by never putting the transfer button on the profile
page. Further, if a GPO user manually navigates to the transfer
page, redirect them to the home page. Finally, if the GPO user
invokes the do-transfer script manually, return an error code.