Update upload-file.php
This one was tricky to fix. This PHP is used by other PHP scripts and the rspec validation is done 3 levels down after the file is uploaded. The bug is triggered during one of these handoffs from php file to file. The right way to fix this would have been to validate the file uploaded right here. I did try to do that using some of the subroutines defined in other scripts, but all of them seem to allow the malicious JS/HTML code to pass thru since they are valid XML/HTML , but just no rspec related stuff in it. SO my quick and dirty fix for this is to load the string into a XML DOM and loo for the "script" TAG and throw an error if one is seen. RSPEC should not have a SCRIPTS tag in it. Period !!!